Shortcut to delete annotations Issue# 9
Hi,
I've created a action Listener named 'KeyPressed' Which will be called when a key is pressed if The key will be 'Delete' it will pass AjaxRequestTarget instance to actionDelete method. I've initialized AjaxRequestTarget delTarget instance in actionSelect() method. Which will save the selected reference in delTarget and send it to the actionDelete() method to delete it.
Regards,
Abdullah Nawaz