Annotation viewer mode
in this pull request, we implemented the feature for Annotation viewer mode. When the feature is enable by ticking the box, it makes the annotation visible to other managers.

As we looked into the feature (Annotation viewer mode), And have found out that the functionality has already been implemented, like the fact that annotations can be viewed by managers. We hit a pause as we discovered that the dialog folder is missing from the source code we were given. This limited our implementation.