Update CasStorageServiceImpl.java
bug because
This goes against the grain when it comes to commenting. In many places, the class has not followed the conventional method for adding comments. That is why non-professional practice is being considered.
 
To solve this bug:
"/.../" should be used for multi-line comments, while "//" should be used for single-line comments. This is a common commenting style. These guidelines should be followed when writing comments.